requirements.txt
streamlit==1.32.0
pandas
SQLAlchemy
psycopg2-binary
python-dotenv